save data

+++ tabibi_system/src/main/java/com/tabibi/tabibi_system/Repositories/PatientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tabibi.tabibi_system.Models.Patient;
 
-public interface PatientRepository extends JpaRepository<Patient,Integer> 
+public interface PatientRepository extends JpaRepository<Patient,Long> 
 {
